マウスホイールの仕様変更 それに伴う拡大縮小の変更

+++ StD/ScrollList.h
@@ -42,7 +42,7 @@ public:
 	void Draw();
 private:
 	// スクロール量
-	int scrollPos_;
+	float scrollPos_;
 	// 描画位置（左上）
 	VECTOR2 pos_;
 	// 描画サイズ
